refactor(repositories): update import paths for UserEntity and EditProfileDto
